Code review changes: Moved generic API functions for reuse, simplified code, adjusted comments
