+++ source/windows/cpu_usage_sampler.c
@@ -9,17 +9,6 @@
  * Base implementation
  ********************************************************************************************************************/
 
-struct aws_cpu_sampler_vtable {
-    int (*aws_get_cpu_sample_fn)(struct aws_cpu_sampler *sampler, double *output);
-    void (*aws_cpu_sampler_destroy)(struct aws_cpu_sampler *sampler);
-};
-
-struct aws_cpu_sampler {
-    const struct aws_cpu_sampler_vtable *vtable;
-    struct aws_allocator *allocator;
-    void *impl;
-};
-
 static void aws_cpu_sampler_destroy_default(struct aws_cpu_sampler *sampler) {
     if (sampler == NULL) {
         return;
@@ -48,26 +37,4 @@ struct aws_cpu_sampler *aws_cpu_sampler_new(struct aws_allocator *allocator) {
     output_unsupported->impl = NULL;
     output_unsupported->vtable = &aws_cpu_sampler_vtable_default;
     return output_unsupported;
-}
-
-void aws_cpu_sampler_destroy(struct aws_cpu_sampler *sampler) {
-    if (sampler == NULL) {
-        return;
-    }
-    if (sampler->vtable->aws_cpu_sampler_destroy == NULL) {
-        return;
-    }
-    sampler->vtable->aws_cpu_sampler_destroy(sampler);
-}
-
-int aws_cpu_sampler_get_sample(struct aws_cpu_sampler *sampler, double *output) {
-    if (sampler == NULL) {
-        aws_raise_error(AWS_ERROR_INVALID_ARGUMENT);
-        return AWS_OP_ERR;
-    }
-    if (sampler->vtable->aws_get_cpu_sample_fn == NULL) {
-        aws_raise_error(AWS_ERROR_PLATFORM_NOT_SUPPORTED);
-        return AWS_OP_ERR;
-    }
-    return sampler->vtable->aws_get_cpu_sample_fn(sampler, output);
 }
